Crate gstreamer_app

Modules

prelude

Structs

AppSink
Appsink is a sink plugin that supports many different methods for making the application get a handle on the GStreamer data in a pipeline. Unlike most GStreamer elements, Appsink provides external API functions.

AppSinkCallbacks

AppSrc
The appsrc element can be used by applications to insert data into a GStreamer pipeline. Unlike most GStreamer elements, appsrc provides external API functions.

AppSrcCallbacks

Continue
Continue calling the closure in the future iterations or drop it.

Error
A generic error capable of representing various error domains (types).

TypedValue
A statically typed Value.

Value
A generic value capable of carrying various types.

Enums

AppStreamType
The stream type.

Type
A GLib or GLib-based library type

Traits

Cast
Upcasting and downcasting support.

IsA
Declares the "is a" relationship.

StaticType
Types that are supported by GLib dynamic typing.

ToValue
Converts to Value.