Struct gstreamer_app::AppSrc

pub struct AppSrc(_, _);

The appsrc element can be used by applications to insert data into a GStreamer pipeline. Unlike most GStreamer elements, appsrc provides external API functions.

appsrc can be used by linking with the libgstapp library to access the methods directly or by using the appsrc action signals.

Before operating appsrc, the caps property must be set to fixed caps describing the format of the data that will be pushed with appsrc. An exception to this is when pushing buffers with unknown caps, in which case no caps should be set. This is typically true of file-like sources that push raw byte buffers. If you don't want to explicitly set the caps, you can use gst_app_src_push_sample. This method gets the caps associated with the sample and sets them on the appsrc replacing any previously set caps (if different from sample's caps).

The main way of handing data to the appsrc element is by calling the AppSrc::push_buffer method or by emitting the push-buffer action signal. This will put the buffer onto a queue from which appsrc will read from in its streaming thread. It is important to note that data transport will not happen from the thread that performed the push-buffer call.

The "max-bytes" property controls how much data can be queued in appsrc before appsrc considers the queue full. A filled internal queue will always signal the "enough-data" signal, which signals the application that it should stop pushing data into appsrc. The "block" property will cause appsrc to block the push-buffer method until free data becomes available again.

When the internal queue is running out of data, the "need-data" signal is emitted, which signals the application that it should start pushing more data into appsrc.

In addition to the "need-data" and "enough-data" signals, appsrc can emit the "seek-data" signal when the "stream-mode" property is set to "seekable" or "random-access". The signal argument will contain the new desired position in the stream expressed in the unit set with the "format" property. After receiving the seek-data signal, the application should push-buffers from the new position.

These signals allow the application to operate the appsrc in two different ways:

The push mode, in which the application repeatedly calls the push-buffer/push-sample method with a new buffer/sample. Optionally, the queue size in the appsrc can be controlled with the enough-data and need-data signals by respectively stopping/starting the push-buffer/push-sample calls. This is a typical mode of operation for the stream-type "stream" and "seekable". Use this mode when implementing various network protocols or hardware devices.

The pull mode, in which the need-data signal triggers the next push-buffer call. This mode is typically used in the "random-access" stream-type. Use this mode for file access or other randomly accessable sources. In this mode, a buffer of exactly the amount of bytes given by the need-data signal should be pushed into appsrc.

In all modes, the size property on appsrc should contain the total stream size in bytes. Setting this property is mandatory in the random-access mode. For the stream and seekable modes, setting this property is optional but recommended.

When the application has finished pushing data into appsrc, it should call AppSrc::end_of_stream or emit the end-of-stream action signal. After this call, no more buffers can be pushed into appsrc until a flushing seek occurs or the state of the appsrc has gone through READY.

Implements

ElementExt, ObjectExt, ObjectExt

Methods

impl AppSrc

fn end_of_stream(&self) -> FlowReturn

Indicates to the appsrc element that the last buffer queued in the element is the last buffer of the stream.

Returns

gst::FlowReturn::Ok when the EOS was successfuly queued. gst::FlowReturn::Flushing when self is not PAUSED or PLAYING.

fn get_caps(&self) -> Option<Caps>

Get the configured caps on self.

Returns

the gst::Caps produced by the source. gst_caps_unref after usage.

fn get_current_level_bytes(&self) -> u64

Get the number of currently queued bytes inside self.

Returns

The number of currently queued bytes.

fn get_duration(&self) -> ClockTime

Get the duration of the stream in nanoseconds. A value of GST_CLOCK_TIME_NONE means that the duration is not known.

Feature: v1_10

Returns

the duration of the stream previously set with AppSrc::set_duration;

fn get_emit_signals(&self) -> bool

Check if appsrc will emit the "new-preroll" and "new-buffer" signals.

Returns

true if self is emitting the "new-preroll" and "new-buffer" signals.

fn get_latency(&self) -> (u64, u64)

Retrieve the min and max latencies in min and max respectively.

min

the min latency

max

the min latency

fn get_max_bytes(&self) -> u64

Get the maximum amount of bytes that can be queued in self.

Returns

The maximum amount of bytes that can be queued.

fn get_size(&self) -> i64

Get the size of the stream in bytes. A value of -1 means that the size is not known.

Returns

the size of the stream previously set with AppSrc::set_size;

fn get_stream_type(&self) -> AppStreamType

Get the stream type. Control the stream type of self with AppSrc::set_stream_type.

Returns

the stream type.

fn push_sample(&self, sample: &Sample) -> FlowReturn

Extract a buffer from the provided sample and adds it to the queue of buffers that the appsrc element will push to its source pad. Any previous caps that were set on appsrc will be replaced by the caps associated with the sample if not equal.

When the block property is TRUE, this function can block until free space becomes available in the queue.

sample

a gst::Sample from which buffer and caps may be extracted

Returns

gst::FlowReturn::Ok when the buffer was successfuly queued. gst::FlowReturn::Flushing when self is not PAUSED or PLAYING. gst::FlowReturn::Eos when EOS occured.

fn set_caps(&self, caps: &Caps)

Set the capabilities on the appsrc element. This function takes a copy of the caps structure. After calling this method, the source will only produce caps that match caps. caps must be fixed and the caps on the buffers must match the caps or left NULL.

caps

caps to set

fn set_duration(&self, duration: ClockTime)

Set the duration of the stream in nanoseconds. A value of GST_CLOCK_TIME_NONE means that the duration is not known.

Feature: v1_10

duration

the duration to set

fn set_emit_signals(&self, emit: bool)

Make appsrc emit the "new-preroll" and "new-buffer" signals. This option is by default disabled because signal emission is expensive and unneeded when the application prefers to operate in pull mode.

emit

the new state

fn set_latency(&self, min: u64, max: u64)

Configure the min and max latency in src. If min is set to -1, the default latency calculations for pseudo-live sources will be used.

min

the min latency

max

the min latency

fn set_max_bytes(&self, max: u64)

Set the maximum amount of bytes that can be queued in self. After the maximum amount of bytes are queued, self will emit the "enough-data" signal.

max

the maximum number of bytes to queue

fn set_size(&self, size: i64)

Set the size of the stream in bytes. A value of -1 means that the size is not known.

size

the size to set

fn set_stream_type(&self, type_: AppStreamType)

Set the stream type on self. For seekable streams, the "seek" signal must be connected to.

A stream_type stream

type_

the new state

fn get_property_block(&self) -> bool

fn set_property_block(&self, block: bool)

fn get_property_current_level_bytes(&self) -> u64

fn get_property_duration(&self) -> u64

fn set_property_duration(&self, duration: u64)

fn get_property_format(&self) -> Format

fn set_property_format(&self, format: Format)

fn get_property_is_live(&self) -> bool

fn set_property_is_live(&self, is_live: bool)

fn get_property_max_latency(&self) -> i64

fn set_property_max_latency(&self, max_latency: i64)

fn get_property_min_latency(&self) -> i64

fn set_property_min_latency(&self, min_latency: i64)

fn get_property_min_percent(&self) -> u32

fn set_property_min_percent(&self, min_percent: u32)

fn connect_enough_data<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_need_data<F: Fn(&AppSrc, u32) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_seek_data<F: Fn(&AppSrc, u64) -> bool + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_block_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_caps_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_current_level_bytes_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_duration_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_emit_signals_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_format_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_is_live_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_max_bytes_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_max_latency_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_min_latency_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_min_percent_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_size_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

fn connect_property_stream_type_notify<F: Fn(&AppSrc) + Send + Sync + 'static>(
&self,
f: F
) -> u64

impl AppSrc

fn push_buffer(&self, buffer: Buffer) -> FlowReturn

Adds a buffer to the queue of buffers that the appsrc element will push to its source pad. This function takes ownership of the buffer.

When the block property is TRUE, this function can block until free space becomes available in the queue.

buffer

a gst::Buffer to push

Returns

gst::FlowReturn::Ok when the buffer was successfuly queued. gst::FlowReturn::Flushing when self is not PAUSED or PLAYING. gst::FlowReturn::Eos when EOS occured.

fn set_callbacks(&self, callbacks: AppSrcCallbacks)

Set callbacks which will be executed when data is needed, enough data has been collected or when a seek should be performed. This is an alternative to using the signals, it has lower overhead and is thus less expensive, but also less flexible.

If callbacks are installed, no signals will be emitted for performance reasons.

callbacks

the callbacks

user_data

a user_data argument for the callbacks

notify

a destroy notify function