gstreamer_app::app_sink

Struct AppSinkBuilder

pub struct AppSinkBuilder<'a> {
    builder: GObjectBuilder<'a, AppSink>,
    callbacks: Option<AppSinkCallbacks>,
    drop_out_of_segment: Option<bool>,
}

A builder-pattern type to construct AppSink objects.

Fields
builder: GObjectBuilder<'a, AppSink>callbacks: Option<AppSinkCallbacks>drop_out_of_segment: Option<bool>
Implementations

impl<'a> AppSinkBuilder<'a>

pub fn build(self) -> AppSink

Build the AppSink.

Panics

This panics if the AppSink doesn’t have all the given properties or property values of the wrong type are provided.

pub fn async_(self, async_: bool) -> Self

pub fn buffer_list(self, buffer_list: bool) -> Self

pub fn callbacks(self, callbacks: AppSinkCallbacks) -> Self

pub fn caps(self, caps: &'a Caps) -> Self

pub fn drop(self, drop: bool) -> Self

pub fn drop_out_of_segment(self, drop_out_of_segment: bool) -> Self

pub fn enable_last_sample(self, enable_last_sample: bool) -> Self

pub fn max_bitrate(self, max_bitrate: u64) -> Self

pub fn max_buffers(self, max_buffers: u32) -> Self

pub fn max_lateness(self, max_lateness: i64) -> Self

pub fn qos(self, qos: bool) -> Self

pub fn render_delay(self, render_delay: Option<ClockTime>) -> Self

pub fn sync(self, sync: bool) -> Self

pub fn throttle_time(self, throttle_time: u64) -> Self

pub fn ts_offset(self, ts_offset: ClockTimeDiff) -> Self

pub fn wait_on_eos(self, wait_on_eos: bool) -> Self

pub fn property(self, name: &'a str, value: impl Into<Value> + 'a) -> Self

Sets property name to the given value value.

Overrides any default or previously defined value for name.

pub fn property_from_str(self, name: &'a str, value: &'a str) -> Self

Sets property name to the given string value value.

pub fn property_if( self, name: &'a str, value: impl Into<Value> + 'a, predicate: bool, ) -> Self

Sets property name to the given inner value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn property_if_some( self, name: &'a str, value: Option<impl Into<Value> + 'a>, ) -> Self

Sets property name to the given inner value if value is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn property_from_iter<V: ValueType + Into<Value> + FromIterator<SendValue>, I: ToSendValue>( self, name: &'a str, iter: impl IntoIterator<Item = I>, ) -> Self

Sets property name using the given ValueType V built from iter’s the Items.

Overrides any default or previously defined value for name.

pub fn property_if_not_empty<V: ValueType + Into<Value> + FromIterator<SendValue>, I: ToSendValue>( self, name: &'a str, iter: impl IntoIterator<Item = I>, ) -> Self

Sets property name using the given ValueType V built from iter’s Items, if iter` is not empty.

This has no effect if iter is empty, i.e. previous value for name is unchanged.

pub fn property_from_str_if( self, name: &'a str, value: &'a str, predicate: bool, ) -> Self

Sets property name to the given string value value if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous value for name is kept.

pub fn property_from_str_if_some( self, name: &'a str, value: Option<&'a str>, ) -> Self

Sets property name to the given string value value if it is Some.

This has no effect if the value is None, i.e. default or previous value for name is kept.

pub fn name(self, name: impl Into<GString>) -> Self

Sets the name property to the given name.

pub fn name_if(self, name: impl Into<GString>, predicate: bool) -> Self

Sets the name property to the given name if the predicate evaluates to true.

This has no effect if the predicate evaluates to false, i.e. default or previous name is kept.

pub fn name_if_some(self, name: Option<impl Into<GString>>) -> Self

Sets the name property to the given name if it is Some.

This has no effect if the value is None, i.e. default or previous name is kept.